requirements.txt
langchain_openai
browser-use
playwright
python-dotenv
